Cor security config

+++ src/main/java/ru/zubov/pizzacloud/controller/rest/IngredientController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import ru.zubov.pizzacloud.entity.Ingredient;
 import ru.zubov.pizzacloud.entity.SignUpDto;
@@ -12,6 +13,7 @@ import ru.zubov.pizzacloud.entity.mapper.IngredientMapper;
 import ru.zubov.pizzacloud.repository.IngredientRepository;
 import ru.zubov.pizzacloud.service.CustomUserDetailsService;
 
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @RestController
@@ -22,6 +24,7 @@ public class IngredientController {
     private final IngredientRepository repo;
     private final IngredientMapper ingredientMapper;
     private final CustomUserDetailsService customUserDetailsService;
+    private final PasswordEncoder passwordEncoder;
 
     @GetMapping
     public Iterable<IngredientDto> allIngredients() {
@@ -42,10 +45,20 @@ public class IngredientController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public ResponseEntity<?> deleteIngredient(@PathVariable("id") String ingredientId, @RequestBody SignUpDto signUpDto) {
         UserDetails userDetails = customUserDetailsService.loadUserByUsername(signUpDto.getLogin());
-        if(userDetails == null){
-            return new ResponseEntity<>("Username is already taken!", HttpStatus.BAD_REQUEST);
+        if (userDetails == null) {
+            return new ResponseEntity<>("Username can't find!", HttpStatus.BAD_REQUEST);
+
         }
-        repo.deleteById(ingredientId);
-        return new ResponseEntity<>(null, HttpStatus.OK);
+
+        if (!Objects.equals(userDetails.getPassword(), passwordEncoder.encode(signUpDto.getPassword()))) {
+            return new ResponseEntity<>("Username or password is wrong!", HttpStatus.BAD_REQUEST);
+        }
+
+        if (userDetails.getAuthorities().stream().anyMatch(a -> Objects.equals(a.getAuthority(), "ROLE_ADMIN"))) {
+            repo.deleteById(ingredientId);
+            return new ResponseEntity<>(null, HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>("Don't have right role!", HttpStatus.BAD_REQUEST);
     }
 }
